Define method for obtaining a label.
I'm going to take it slowly so that you can see my thought process.

This is something that a carrier is "asked" to do, so the method ought to be defined on the Carrier superclass.  It is sensible to take method parameters exactly like `find_rates` for the sake of consistency.