Extend label integration test, get proper label.
The second phase (accept) is where the label image comes from.  To get the image, you use the digest from the first phase (confirm).  This let's you verify cost information and other such details presumably.

I still need to handle API validation errors and figure out how to expose the data retrieved from UPS.

Furthermore, the UPS API supports a variety of options for shipping that I must be sure to support in order for this to be useful to anyone other than me.