Preliminary Shipper/ShipFrom support.
The Shipper (return address on label) sometimes needs to be different from the ShipFrom (pickup location used for calculating rates).  For example, a company that has a contractor build whatever needs to be shipped.

UPS requires a company name when a ShipFrom element is present, whereas it does not require this for the Shipper element.  I need to figure out if and how to make this obvious to my fellow developers.